+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Wed Mar 12 18:04:28 GMT 2014 : menu links now direct to resources
 - Wed Mar 12 18:08:33 GMT 2014 : showImage is not a function error
 - Wed Mar 12 18:41:45 GMT 2014 : spinner showing up
+- Wed Mar 12 19:26:11 GMT 2014 : image summary is now on the bottom